add type

+++ blog-dao/src/main/java/com/zs/blog/model/Role.java
@@ -38,7 +38,7 @@ public class Role implements Serializable {
     /**
      * 1:正常,2:删除
      */
-    private Byte state;
+    private Integer state;
 
     /**
      * 创建时间
